api_service/requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
pydantic==2.10.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
passlib==1.7.4
pyjwt==2.10.1
bcrypt==4.2.1
requests==2.32.3
dramatiq==1.17.1
redis==5.2.1
apscheduler==3.11.0
Jinja2==3.1.5
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.25.2
mypy==1.14.1
black
isort
